added a controller
