+++ src/routers/todo.js
@@ -1,5 +1,5 @@
 const express = require('express');
-
+const controller = require('../controllers/controller');
 //importamos el fichero con los datos que necesita nuestro Router
 const {todos} = require('../data/index');
 
@@ -16,29 +16,12 @@ const todoRouter = express.Router();
 
  //devolver todos los "to-dos" que hay en el array con formato JSON.
   //el codigo de status 200 significa que la peticion se ha procesado correctamente.
-todoRouter.get('/todo', (req, res) => {
-  res.status(200).json(todos);
-});
-
+todoRouter.get('/', controller.getTodo);
 
   //crear un nuevo objeto con estructura {id, text, fecha, done} con los datos que vienen en el BODY de la Request y meterlos dentro de el array.
   //el nuevo objeto debe tener como id un numero mas que el numero actual de elementos guardados en el array.
 
-  todoRouter.post('/todo', (req, res) => {
-    if (!req.body.title || typeof req.body.title !== 'string') {
-      return res.status(400).json({ error: 'Title is required and must be a string' });
-    }
-
-    const newTodo = {
-      id: todos.length,
-      title: req.body.title,
-      fecha: req.body.fecha,
-      done: req.body.done
-    };
-
-    todos.push(newTodo);
-    res.status(201).json(newTodo);
-  });
+todoRouter.post('/', controller.postTodo);
 
 /*
 En este endpoint, el path contiene una variable llamada id. La syntaxis que utiliza express para estos casos es el simbolo :
@@ -53,17 +36,7 @@ Si con Insomnia o Postman hicisemos una peticion GET a la ruta /todo/12, está s
 
 */
 
-todoRouter.get('/todo/:id',  (req, res) => {
-
-  const todoFound = todos.find(t => t.id === Number(req.params.id));
-
- if(!todoFound) {
-  res.status(404).json({error: 'To-do not found'});
- } else {
-  res.status(200).json(todoFound);
-//  }
-}
-});
+todoRouter.get('/:id', controller.getTodoById);
 
 
   //recogemos el valor de la variable del path llamada "id" y lo transformarlo a un numero (todos nuestros ids son numericos).
@@ -77,19 +50,7 @@ todoRouter.get('/todo/:id',  (req, res) => {
 
 // MISSING '/todo/:id' PATCH
 
-todoRouter.patch('/todo/:id', (req, res) => {
-  const todoIndex = todos.findIndex(t => t.id === Number(req.params.id));
-
-  if (!todoIndex) {
-    return res.status(404).json({ error: 'To-do not found' });
-  }
-
-  const updatedTodo = { ...todos[todoIndex], ...req.body };
-  todos[todoIndex] = updatedTodo;
-
-  return res.json(updatedTodo);
-});
-
+todoRouter.patch('/:id', controller.updateTodo);
 
 
 //recogemos el valor de la variable del path llamada "id" y lo transformarlo a un numero (todos nuestros ids son numericos).
@@ -107,17 +68,7 @@ todoRouter.patch('/todo/:id', (req, res) => {
 // MISSING '/todo/:id' DELETE
 
 
-todoRouter.delete('/todo/:id',  (req, res) => {
-  const todoIndex = todos.findIndex(t => t.id === Number(req.params.id));
-
-  if (todoIndex === -1) {
-    return res.status(404).json({ error: 'To-do not found' });
-  }
-
-  todos.splice(todoIndex, 1);
-
-  return res.status(204).send();
-});
+todoRouter.delete('/:id', controller.deleteTodo);
 
   //recogemos el valor de la variable del path llamada "id" y lo transformarlo a un numero (todos nuestros ids son numericos).
   //cualquier valor que recogemos de req.params será siempre un String. Por eso lo debemos convertir a numero.
